+++ src/elements/dest.c
@@ -45,9 +45,9 @@ int update_DEST(UPDATE_FUNC_ARGS) {
 	{
 		if (ptypes[r&0xFF].hconduct) parts[r>>8].temp = restrict_flt(parts[r>>8].temp+10000.0f, MIN_TEMP, MAX_TEMP);
 	}
-	topv=pv[y/CELL][x/CELL]/9+parts[r>>8].temp/900;
+	topv=(int)(pv[y/CELL][x/CELL]/9+parts[r>>8].temp/900);
 	if (topv>40.0f)
-		topv=40.0f;
+		topv=40;
 	pv[y/CELL][x/CELL]+=40.0f+topv;
 	parts[i].temp = restrict_flt(parts[i].temp, MIN_TEMP, MAX_TEMP);
 	return 0;
